Refactor stock fetching to use BigDecimal and parameterize symbol
- Change the `price` field in `StockSnapshot` from `double` to `BigDecimal` for better precision.
- Update `StockFetcher.fetch` to accept a `symbol` parameter instead of hardcoding the index.
- Modify `StockFetcher` to extract and return the price as a `BigDecimal` directly.
- Introduce a `DOW_JONES_SYMBOL` constant in `App` and pass it to the fetcher.
- Increase the scheduled task interval in `App` from 5 to 60 seconds.

+++ app/src/main/java/org/example/model/StockSnapshot.java
@@ -1,8 +1,9 @@
 package org.example.model;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record StockSnapshot(
         Instant timestamp,
-        double price
+        BigDecimal price
 ) {}
